server/requirements.txt
torch>=2.1.0
git+https://github.com/huggingface/diffusers
transformers
accelerate
safetensors
sentencepiece
Pillow
fastapi
uvicorn[standard]
python-multipart
